# docs/swagger.yaml
basePath: /
definitions:
  github_com_telkomdev-candrasaka_hris-monitoring_git_internal_repositories.AttendanceRiskReportRow:
    properties:
      approved_leave_count:
        type: integer
      attendance_records:
        type: integer
      attendance_risk_score:
        type: number
      city:
        type: string
      estimated_available_staff:
        type: integer
      headcount:
        type: integer
      location_id:
        type: integer
      location_name:
        type: string
      location_type:
        type: string
      minimum_staffing:
        type: integer
      open_attendance_count:
        type: integer
      pending_leave_count:
        type: integer
      staffing_risk:
        type: string
    type: object
  github_com_telkomdev-candrasaka_hris-monitoring_git_internal_repositories.LabourCostLeakageReportRow:
    properties:
      city:
        type: string
      headcount:
        type: integer
      leakage_percentage:
        type: number
      location_id:
        type: integer
      location_name:
        type: string
      location_type:
        type: string
      total_absent_count:
        type: integer
      total_base_salary:
        type: number
      total_deduction:
        type: number
      total_gross_salary:
        type: number
      total_late_count:
        type: integer
      total_net_salary:
        type: number
    type: object
...
  internal_handlers.AttendanceRiskReportListDoc:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_telkomdev-candrasaka_hris-monitoring_git_internal_repositories.AttendanceRiskReportRow'
        type: array
    type: object
...
  internal_handlers.CreateUserRequestDoc:
    properties:
      base_salary:
        example: 5000000
        type: number
      email:
        example: budi@example.com
        type: string
      location_id:
        example: 1
        type: integer
      name:
        example: Budi
        type: string
      password:
        example: secret123
        type: string
      role:
        example: staff_gudang
        type: string
      shift_id:
        example: 2
        type: integer
    type: object
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_handlers.LabourCostLeakageReportListDoc:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_telkomdev-candrasaka_hris-monitoring_git_internal_repositories.LabourCostLeakageReportRow'
        type: array
    type: object
...
  internal_handlers.LocationResponseDoc:
    properties:
      address:
        example: Jl. Raya Bekasi No. 1
        type: string
      city:
        example: Bekasi
        type: string
      geofence_radius:
        example: 100
        type: number
      id:
        example: 1
        type: integer
      latitude:
        example: -6.2
        type: number
      longitude:
        example: 106.9
        type: number
      minimum_staffing:
        example: 8
        type: integer
      name:
        example: Warehouse Bekasi
        type: string
      province:
        example: Jawa Barat
        type: string
      type:
        example: warehouse
        type: string
    type: object
...
  internal_handlers.ShiftResponseDoc:
    properties:
      cross_midnight:
        example: false
        type: boolean
      end_time:
        example: "17:00"
        type: string
      grace_minutes:
        example: 15
        type: integer
      id:
        example: 2
        type: integer
      location_id:
        example: 1
        type: integer
      name:
        example: Shift Pagi
        type: string
      start_time:
        example: "08:00"
        type: string
    type: object
...
  internal_handlers.UpdateUserRequestDoc:
    properties:
      base_salary:
        example: 6500000
        type: number
      email:
        example: budi.updated@example.com
        type: string
      location_id:
        example: 1
        type: integer
      name:
        example: Budi Update
        type: string
      password:
        example: newsecret123
        type: string
      role:
        example: manager_outlet
        type: string
      shift_id:
        example: 2
        type: integer
    type: object
  internal_handlers.UserResponseDoc:
    properties:
      base_salary:
        example: 5000000
        type: number
      email:
        example: budi@example.com
        type: string
      id:
        example: 10
        type: integer
      location:
        $ref: '#/definitions/internal_handlers.LocationResponseDoc'
      location_id:
        example: 1
        type: integer
      name:
        example: Budi
        type: string
      role:
        example: staff_gudang
        type: string
      shift:
        $ref: '#/definitions/internal_handlers.ShiftResponseDoc'
      shift_id:
        example: 2
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend HRIS dan operasional frozen-food retail dengan JWT auth, geofenced
    attendance + selfie, leave approval with staffing risk, payroll, PPE/asset compliance,
    dan executive reports.
  title: HRIS Monitoring API
  version: "1.0"
paths:
  /api/reports/attendance-risk:
    get:
      description: Executive report showing attendance anomalies, leave load, minimum
        staffing pressure, and computed attendance risk by location.
      parameters:
      - description: Month
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.AttendanceRiskReportListDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get attendance risk report
      tags:
      - Reports
  /api/reports/labour-cost-leakage:
    get:
      description: Executive report showing attendance-related payroll leakage by
        location, including deductions, lateness, absence counts, and leakage percentage.
      parameters:
      - description: Month
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.LabourCostLeakageReportListDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get labour cost leakage report
      tags:
      - Reports
...
  /users:
    get:
      description: List employees and managers with their assigned location, shift,
        and base salary.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.UserResponseDoc'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user account and assign role, location, optional shift,
        and base salary.
      parameters:
      - description: User payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateUserRequestDoc'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.UserResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Delete a user. Deletion is blocked if the user still holds borrowed
        assets.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      description: Retrieve one user by identifier.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.UserResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update mutable user fields such as role, location, shift, password,
        and salary.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update user payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateUserRequestDoc'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.UserResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"